parallel_catchup/requirements.txt
requests==2.32.3
kubernetes==30.1.0
PyYAML==6.0.1